requirements.txt
yfinance>=0.2.40
google-genai>=0.3.0
